Renamed packages' names
