--- utils/file_utils/doc.go
@@ -1,2 +1,0 @@
-// Package file_utils contains utilities that might be useful in working with files.
-package file_utils
