+++ utils/fileutils/doc.go
@@ -1,0 +1,2 @@
+// Package fileutils contains utilities that might be useful in working with files.
+package fileutils
