+++ utils/fileutils/testdata/copy2.input
