+++ utils/fileutils/testdata/copy2.output
