+++ utils/fileutils/testdata/copy3.input
