+++ utils/fileutils/testdata/copy3.output
