+++ utils/fileutils/testdata/remove_tabs2.input
