+++ utils/fileutils/testdata/remove_tabs3.output
