+++ utils/fileutils/testdata/remove_tabs4.input
