--- utils/number_utils/doc.go
@@ -1,2 +1,0 @@
-// Package number_utils contains utilities that might be useful in working with numbers.
-package number_utils
