+++ utils/numutils/doc.go
@@ -1,0 +1,2 @@
+// Package numutils contains utilities that might be useful in working with numbers.
+package numutils
