+++ utils/numutils/fibonacci.go
@@ -1,4 +1,4 @@
-package number_utils
+package numutils
 
 import "fmt"
