+++ utils/numutils/is_odd.go
@@ -1,4 +1,4 @@
-package number_utils
+package numutils
 
 // IsOdd takes an Integer and checks if it's odd.
 func IsOdd(x int) (r bool) {
