+++ utils/numutils/is_odd_test.go
@@ -1,4 +1,4 @@
-package number_utils
+package numutils
 
 import "testing"
