+++ utils/numutils/square_root_test.go
@@ -1,4 +1,4 @@
-package number_utils
+package numutils
 
 import (
 	"math"
