--- utils/string_utils/doc.go
@@ -1,2 +1,0 @@
-// Package string_utils contains utilities that might be useful in working with strings.
-package string_utils
