+++ utils/strutils/doc.go
@@ -1,0 +1,2 @@
+// Package strutils contains utilities that might be useful in working with strings.
+package strutils
