+++ utils/strutils/rotate13.go
@@ -1,4 +1,4 @@
-package string_utils
+package strutils
 
 import (
 	"bytes"
