+++ utils/strutils/to_upper_case.go
@@ -1,4 +1,4 @@
-package string_utils
+package strutils
 
 // ToUpperCase takes string and returns its copy transformed to contain only upper-letters.
 func ToUpperCase(s string) string {
