+++ utils/strutils/to_upper_case_test.go
@@ -1,4 +1,4 @@
-package string_utils
+package strutils
 
 import "testing"
